修改服务员销售统计表存储过程

+++ newspicyway/sql_script/9-分店-procedure-report.sql
@@ -8317,7 +8317,7 @@ BEGIN
     superkey varchar(50),
     dishnum varchar(50),
     dishtype int,
-    orderprice double(8, 2),
+    orderprice DECIMAL(10, 2),
     dishunit varchar(100)
   ) ENGINE = MEMORY DEFAULT charset = utf8;
